change styles + routing
