+++ src/components/orders-list-item/orders-list-item.module.scss
@@ -10,6 +10,7 @@
 }
 
 .badges {
+	display: flex;
 	width: 100%;
 	height: 64px;
 	z-index: 0;
