+++ src/pages/profile-page/profile-page.module.scss
@@ -1,17 +1,12 @@
-.profileWrapper {
-	width: 100%;
-	height: 100%;
-	box-sizing: border-box;
-	max-width: 1320px;
-	margin-left: auto;
-	margin-right: auto;
+.layout {
 	display: flex;
-	gap: 60px;
+	width: 100%;
+	padding: 0 20px;
+	gap: 80px;
 }
 
-.navWrapper {
-	display: flex;
-	justify-content: space-between;
+.outlet {
+	max-height: calc(100vh - 140px);
 }
 
 .link {
@@ -32,10 +27,4 @@
 .linkActive {
 	color: #f2f2f3;
 	text-decoration: none;
-}
-
-.profileForm {
-	display: flex;
-	flex-direction: column;
-	gap: 20px;
 }
