+++ src/styles.css
@@ -41,6 +41,11 @@ body {
 	margin: 0;
 }
 
+main {
+	max-width: 1300px;
+	margin: 0 auto;
+}
+
 h1 {
 	font-size: 3.2em;
 	line-height: 1.1;
